% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diffOdsBill_query}
\alias{diffOdsBill_query}
\title{ODS表的差异检查}
\usage{
diffOdsBill_query(token = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{token}{口令}
}
\value{
返回值
}
\description{
ODS表的差异检查
}
\examples{
diffOdsBill_query()
}
