% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_salary.R
\name{ds_salaryOdsCheckBank_query}
\alias{ds_salaryOdsCheckBank_query}
\title{工资ODS表异常检查}
\usage{
ds_salaryOdsCheckBank_query(FToken = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{FToken}{口令}
}
\value{
返回值
}
\description{
工资ODS表异常检查
}
\examples{
voucherRuleCheck_query()
}
