% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_socialSecury.R
\name{ds_socialSecuryOdsCheck_deal}
\alias{ds_socialSecuryOdsCheck_deal}
\title{社保ODS表异常处理}
\usage{
ds_socialSecuryOdsCheck_deal(FToken = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{FToken}{口令}
}
\value{
返回值
}
\description{
社保ODS表异常处理
}
\examples{
voucherRuleCheck_deal()
}
