% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{odsBillList_Query}
\alias{odsBillList_Query}
\title{单据对象查询}
\usage{
odsBillList_Query(token = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{token}{口令}
}
\value{
返回值
}
\description{
单据对象查询
}
\examples{
odsBillList_Query()
}
