% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_voucher.R
\name{rule_voucherCheck_deal}
\alias{rule_voucherCheck_deal}
\title{凭证规则表异常处理}
\usage{
rule_voucherCheck_deal(FToken = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{FToken}{口令}
}
\value{
返回值
}
\description{
凭证规则表异常处理
}
\examples{
voucherRuleCheck_deal()
}
