% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl.R
\name{tpl_rdCumFund_add}
\alias{tpl_rdCumFund_add}
\title{插入研发支出的公积金数据}
\usage{
tpl_rdCumFund_add(
  token = "057A7F0E-F187-4975-8873-AF71666429AB",
  FNumber = "B018",
  FSeq_ref = 2,
  FSeq_new = 3
)
}
\arguments{
\item{token}{口令}

\item{FNumber}{模板编号}

\item{FSeq_ref}{模板序号}

\item{FSeq_new}{模板新行事收}
}
\value{
返回值
}
\description{
插入研发支出的公积金数据
}
\examples{
tpl_rdSocialSecurity_add()
}
