% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl.R
\name{tpl_socialSecurityDisability_add}
\alias{tpl_socialSecurityDisability_add}
\title{插入管理费用-社保}
\usage{
tpl_socialSecurityDisability_add(
  token = "057A7F0E-F187-4975-8873-AF71666429AB",
  FNumber = "B018",
  FSeq_ref = 2,
  FSeq_new = 3
)
}
\arguments{
\item{token}{口令}

\item{FNumber}{模板编号}

\item{FSeq_ref}{模板序号}

\item{FSeq_new}{模板新行事收}
}
\value{
返回值
}
\description{
插入管理费用-社保
}
\examples{
tpl_socialSecurityDisability_add()
}
